Auto_LLM_Evaluator/requirements.txt
pandas>=1.5
numpy>=1.26.4, <2
regex>=2023.10.3
PyYAML>=6.0
google-cloud-aiplatform==1.76.0
sentence-transformers>=4.0.1

